Update attribute xml to match with fapi2
There were a lot of attributes whose type did not match what was in
the EKB. This commit fixes a lot of those issues

RTC: 158587
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/28634
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/usr/targeting/attrsync.C
@@ -678,7 +678,7 @@ namespace TARGETING
                 for(TargetHandleList::const_iterator mca = l_mcaList.begin();
                     mca != l_mcaList.end(); ++mca)
                 {
-                    uint8_t effDimmSize[][2] = {{0x4,0x0},{0x4,0x0}};
+                    uint32_t effDimmSize[][2] = {{0x4,0x0},{0x4,0x0}};
                     (*mca)->setAttr<ATTR_EFF_DIMM_SIZE>(effDimmSize);
                 }
